DOC: Build versioned docs and deploy them to gh-pages
This PR adds three new build steps to our CircleCI workflow.
The first one, called ``build_docs`` generates the documentation
of the project using sphinx.
In the case the commit message contains the tag ``[docs versions]``
(or a few variations thereof), then a readthedocs-style of versioned
documentation is built.
The other two build steps are run only for tag-commits and the master
branch, to deploy the new versioned documentation to GitHub pages.

I've set up the corresponding write-access keys in the repo and
CircleCI, so this first step towards having documentation should fairly
good to go.