chore: remove the leftovers of AFQ in ``.afq/`` [skip ci]
After all we keep the pre-nipreps branch, and it is distorting a bit the
line counts when updating the ``.zenodo.json`` file.

Hoping there's no conflict with this from @richford's point of view.

The new `.zenodo.json` file shows how @josephmje would be now placed at
the top.

In addition, here I add @mattcieslak to the creators list as his name
has been missing for a little while. Judging by the extremely limited
functionality at this point, and the zero papers about dMRIPrep I hope
that has not been an issue.